Changed datatype of deptID

+++ src/com/java/Department.java
@@ -5,7 +5,7 @@ public class Department {
 public static void main(String[] args) {
 
 	String deptName;
-	int deptID;
+	double deptID;
 	String Section;
 	//change3
 	//change4
